% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_selectivity.R
\name{analyze_selectivity}
\alias{analyze_selectivity}
\title{Quantify sample selectivity with reduced/cleaned data}
\usage{
analyze_selectivity(.base_sample, .analysis_sample, .cols)
}
\arguments{
\item{.base_sample}{A data frame. Should contain the data of the
total base sample.}

\item{.analysis_sample}{A data frame. Should contain the data of the
reduced/cleaned analysis sample (e.g., after excluding observations).
The number of observations in the analysis sample should be
smaller than in the total base sample.}

\item{.cols}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
The names or column numbers of the numeric variables
to compute Cohen's d for. Both \code{.base_sample} and \code{.analysis_sample}
must contain the columns specified.}
}
\value{
A tibble with 6 columns:
\itemize{
\item{\code{variable}} {The column the sensitivity anlysis was performed for}
\item{\code{mean_tbs}} {Mean in the total base sample}
\item{\code{sd_tbs}} {Standard deviation in the total base sample}
\item{\code{mean_as}} {Mean in the reduced/cleaned analysis sample}
\item{\code{sd_as}} {Standard deviation in the reduced/cleaned analysis sample}
\item{\code{d}} {Cohen's d quantifying the sample selectivity}
}
}
\description{
This function performs sensitivity analysis by
computing the effect size Cohen's d per column
to quantify sample selectivity of a reduced/cleaned (i.e., selective)
analysis sample (e.g., after excluding observations)
compared to its total base sample.
}
\details{
Cohen's d is calculated with
\eqn{( M(reduced) - M(total) ) / SD(total)}.
}
\examples{
\dontrun{
analyze_selectivity(.base_sample = studach,
.analysis_sample = studach_reduced, .cols = gender:read)
}
}
