% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_means.R
\name{cluster_means}
\alias{cluster_means}
\title{Compute cluster means}
\usage{
cluster_means(data, .cols, .id_cluster, .suffix = "_cl")
}
\arguments{
\item{data}{A data frame.}

\item{.cols}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
The name/s or column number/s of the numeric variable/s for which
cluster means should be computed.}

\item{.id_cluster}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
The name or column number of the identifier for the clusters within which
means should be computed.}

\item{.suffix}{Optional. A string to be added to the original variable
name/s to label the new column/s for the cluster means. The default is the
suffix "_cl".}
}
\value{
An tibble containing the new cluster means indicated by the suffix
specified in \code{.suffix}.
}
\description{
This function computes the mean of a single or a range of
specified variable/s within each cluster.
}
\examples{
# Compute cluster means at the school level for mathematics
cluster_means(data = studach, .cols = math, .id_cluster = id_sch, .suffix = "_sch")

# Compute cluster means at the classroom level for mathematics and reading
cluster_means(data = studach, .cols = math:read, .id_cluster = id_cla, .suffix = "_cla")
}
