% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose}
\alias{compose}
\title{Create a composite measure}
\usage{
compose(data, .cols, .composite = "composite", .scale = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{.cols}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
The names or column numbers of the numeric variables to be integrated into
the composite measure.}

\item{.composite}{The name of the new column for the composite measure.
If no name is supplied, the new composite measure will be named "composite".}

\item{.scale}{A logical indicating whether columns should be scaled in
advance (i.e., by means of z standardization). The default is TRUE.}
}
\value{
An tibble containing the new composite measure named as specified in
\code{.composite}.
}
\description{
This function integrates a range of specified variables
into a simple composite measure
by computing the mean for each observation.
By default, the respective variables are standardized in advance.
}
\examples{
compose(data = studach, .cols = math:read, .composite = "achievement")
}
