% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate_ml.R
\name{correlate_ml}
\alias{correlate_ml}
\title{Compute single- or multilevel correlations}
\usage{
correlate_ml(data, .cols, .id_cluster)
}
\arguments{
\item{data}{A data frame.}

\item{.cols}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
The names or column numbers of the numeric variables
to compute correlations for.}

\item{.id_cluster}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
Optional. The name or column number of the cluster identifier
to compute multilevel correlations.}
}
\value{
A tibble containing a correlation matrix of variables specified in \code{.cols}.
}
\description{
This function computes correlations between a series of specified variables.
If no cluster identifier is specified, the function provides simple
single-level correlations at the individual level of observations.
If a cluster identifier is specified, the function provides
multilevel correlations between the respective group means
at the specified hierarchical level.
}
\examples{
# compute single-level correlations between students' mathematics and reading achievement
correlate_ml(data = studach, .cols = read:math)

# compute multilevel correlations between students' mathematics and reading achievement
# at the school level
correlate_ml(data = studach, .cols = c(gender, read, math), .id_cluster = id_sch)
}
