% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xlsx_overview.R
\name{create_xlsx_overview}
\alias{create_xlsx_overview}
\title{Create an Excel workbook providing an overview with descriptive statistics and correlations}
\usage{
create_xlsx_overview(
  data,
  .cols,
  .file = "overview.xlsx",
  .group = NULL,
  .id_cluster_l2 = NULL,
  .id_cluster_l3 = NULL,
  .round = 3
)
}
\arguments{
\item{data}{A data frame.}

\item{.cols}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
The names or column numbers of the numeric variables
to create the overview for.}

\item{.file}{A character string specifying the file name and the path
where to save the .xlsx file.
The default is to save an .xlsx file named 'overview' in the current working directory.}

\item{.group}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
Optional. The name or column number of the grouping variable
to compute descriptive statistics by group. The default is NULL.}

\item{.id_cluster_l2}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
Optional. The name or column number of the identifier for the second
hierarchical level (e.g., classrooms). The default is NULL.}

\item{.id_cluster_l3}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
Optional. The name or column number of the identifier for the third
hierarchical level (e.g., schools). The default is NULL.}

\item{.round}{Optional. An integer constant indicating the number of digits
for rounding the results. The default is to print 3 digits.}
}
\value{
An Excel workbook (as .xlsx file) saved in \code{.file} with multiple sheets
providing an overview with descriptive statistics and correlations
for the variables specified in \code{.cols}.
}
\description{
This function uses the \strong{openxlsx} package to create an Excel workbook
with multiple sheets that contain descriptive statistics and
correlations for a specified variable set.
There are options to specify a) a grouping variable to include descriptive statistics that
are calculated separately by group and b) up to two cluster identifiers
to compute multilevel correlations between the respective group means
at the specified hierarchical levels.
}
\examples{
\dontrun{
create_xlsx_overview(data = studach, .cols = gender:read,
.file = "student_achievement_overview.xlsx",
.group = ts, .id_cluster_l2 = id_cla, .id_cluster_l3 = id_sch,
.round = 3)
}
}
\seealso{
See the \strong{openxlsx} package.
}
