% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_stats.R
\name{describe_stats}
\alias{describe_stats}
\title{Compute descriptive statistics}
\usage{
describe_stats(data, .cols, .group = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{.cols}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
The name/s or column number/s of the
numeric variable/s to compute descriptive statistics for.}

\item{.group}{<\code{\link[dplyr:dplyr_data_masking]{data-masked}}>
Optional. The name of the grouping variable. The default is NULL.}
}
\value{
A tibble or a tibble grouped by \code{.group} containing
the following statistics:
\itemize{
\item{\code{n}} {Number of valid observations}
\item{\code{missings}} {Percentage of missings}
\item{\code{mean}} {Mean}
\item{\code{sd}} {Standard deviation}
\item{\code{min}} {Minimum}
\item{\code{p25}} {25th percentile}
\item{\code{median}} {Median}
\item{\code{p75}} {75th percentile}
\item{\code{max}} {Maximum}
\item{\code{mean_h}} {Harmonic mean}
\item{\code{skewness}} {Skewness}
\item{\code{kurtosis}} {Kurtosis}
}
}
\description{
This function provides a range of descriptive statistics
that are useful to explore numeric data.
If a grouping variable is specified, the descriptive statistics are
computed separately for each group.
}
\examples{
# -- Simple descriptive statistics across all observations

# compute descriptive statistics for students' mathematics achievement

# using the name of the variable
describe_stats(studach, .cols = math)

# using the position by column number
describe_stats(studach, .cols = 11)

# compute descriptive statistics for all numeric variables
describe_stats(studach, .cols = where(is.numeric))

# -- Descriptive statistics by school type or gender
describe_stats(studach, .cols = math, .group = ts_name)
describe_stats(studach, .cols = math, .group = gender)

}
