% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\alias{r2.delta}
\title{Compute (delta) squared multiple correlation coefficients with variances}
\usage{
r2(var, var_adj, N = NULL, trunc = c("zero", "na", "none"))

r2.delta(
  r2_full,
  r2_reduced,
  N = NULL,
  trunc = c("zero", "na", "none"),
  var_delta0 = FALSE
)
}
\arguments{
\item{var}{A numeric vector of the unconditional (i.e., unadjusted)
variance.}

\item{var_adj}{A numeric vector of the conditional (i.e., covariate-adjusted)
variance.}

\item{N}{Optional. A numeric vector of the sample size. Necessary for
computing the variance.}

\item{trunc}{A string determining the treatment of negative
(\eqn{\Delta})\eqn{R^2} values. The default is "zero".
\itemize{
\item "zero": Negative values are truncated to zero.
\item "na": Negative values are set to NA.
\item "none": Negative values are kept. Note that no variances are computed
for negative (\eqn{\Delta})\eqn{R^2} values.
}}

\item{r2_full}{A numeric vector of \eqn{R^2} for a model A with a set of
covariates.}

\item{r2_reduced}{A numeric vector of \eqn{R^2} for a model B with a subset
of the covariates from model A.}

\item{var_delta0}{A logical indicating whether variances of zero
\eqn{\Delta R^2} should be computed. The default is FALSE.}
}
\value{
A matrix that contains:
\itemize{
\item \code{r2()}
\itemize{
\item \code{Estimate} \eqn{R^2}.
\item \code{Variance} Variance of \eqn{R^2}, if \code{N} is supplied.
}
\item \code{r2.delta()}
\itemize{
\item \code{Estimate} \eqn{\Delta}\eqn{R^2}.
\item \code{Variance} Variance of \eqn{\Delta}\eqn{R^2}, if \code{N} is supplied.
}
}
}
\description{
These functions compute squared multiple correlation
coefficients (i.e., proportions of explained variance) \eqn{R^2} for
(a set of) covariates based on (un)conditional variances and
differences \eqn{\Delta} between the \eqn{R^2} values of two nested covariate
models A and B (where the covariates in model B represent a subset of the
covariates in model A), as well as corresponding large-sample variances
implementing the formulas given in Hedges & Hedberg (2013) and
Alf & Graf (1999), respectively.

\itemize{
\item \code{r2()} computes \eqn{R^2} and its variance
(Hedges & Hedberg, 2013, p. 451).
\item \code{r2.delta()} computes \eqn{\Delta R^2} and its variance
(Alf & Graf, 1999, Equations (19) and (21)).
}
}
\details{
\strong{Squared multiple correlation coefficients}

Covariates that are correlated with (or, put differently: explain variance
in) an outcome reduce error variance. \eqn{R^2} quantifies the proportion of
the total variance in an outcome that can be explained by one or more
covariates. Formally, \eqn{R^2} is the ratio of the difference between the
unconditional variance and the on covariates \eqn{C} conditional variance
to the unconditional variance (\eqn{(\sigma^2-\sigma^2_{|C})/\sigma^2}).

\strong{Multilevel designs}

In multilevel models, covariates may act at either hierarchical level.
Generally, group-mean centering of within-cluster covariates is
recommended to ensure that covariates explain variance exclusively at that
hierarchical level at which they are specified
(e.g., Konstantopoulos, 2012). Consider, for example, a three-level design,
where students at L1 are nested within classrooms at L2, which are in turn
nested within schools at L3.
Given (a set of) classroom-mean centered L1 covariates \eqn{C_{L1}},
\eqn{R^2_{L1}} is the ratio of the difference between the unconditional and
the conditional between-student-within-classroom variances to the
unconditional between-student-within-classroom variance located at L1
(\eqn{(\sigma^2_{L1}-\sigma^2_{L1|C_{L1}})/\sigma^2_{L1}});
given (a set of) school-mean centered L2 covariates \eqn{C_{L2}},
\eqn{R^2_{L2}} is the ratio of the difference between the unconditional and
the conditional between-classroom-within-school variances to the
unconditional between-classroom-within-school variance located at L2
(\eqn{(\sigma^2_{L2}-\sigma^2_{L2|C_{L2}})/\sigma^2_{L2}}); and finally,
given (a set of) L3 covariates \eqn{C_{L3}},
\eqn{R^2_{L3}} is the ratio of the difference between the unconditional and
the conditional between-school variances to the
unconditional between-school variance located at L3
(\eqn{(\sigma^2_{L3}-\sigma^2_{L3|C_{L3}})/\sigma^2_{L3}}).

\strong{Delta squared multiple correlation coefficients}

\eqn{\Delta R^2} quantifies the increment in the proportion of explained
variance when adding covariates to a model. Formally, \eqn{\Delta R^2} is the
difference between the \eqn{R^2} of a full model A with a certain set of
covariates and the \eqn{R^2} of a reduced model B with a subset of those
covariates that are included in model A (\eqn{R^2_A-R^2_B}).
}
\examples{
# compute R-squared and its sampling variance
r2(var = 3.25, var_adj = 1.75, N = 100, trunc = "zero")

# compute Delta R-squared and its sampling variance
r2.delta(r2_full = .75, r2_reduced = .50, N = 100, trunc = "zero", var_delta0 = FALSE)


}
\references{
Alf, E. F., & Graf, R. G. (1999). Asymptotic confidence limits for the
difference between two squared multiple correlations: A simplified approach.
\emph{Psychological Methods}, \emph{4}(1), 70–75.
https://doi.org/10.1037/1082-989X.4.1.70

Hedges, L. V., & Hedberg, E. C. (2013). Intraclass correlations and
covariate outcome correlations for planning two- and three-level
cluster-randomized experiments in education. \emph{Evaluation Review},
\emph{37}(6), 445–489. https://doi.org/10.1177/0193841X14529126

Konstantopoulos, S. (2012). The impact of covariates on statistical power
in cluster randomized designs: Which level matters more?
\emph{Multivariate Behavioral Research}, \emph{47}(3), 392–420.
https://doi.org/10.1080/00273171.2012.673898
}
